ARM: dts: loire: Change default state for gpio48 to output-high
This pin needs to be in a high state to allow proper
detection of the wifi card at sdhci_3.

Doing this in the drivers, that need this pin in a
high state to probe and work, has been shown to be
pretty ugly.

Hence, changing the default state was deemed to be
a better solution.

Change-Id: I0d609e2644cec318b88cb2075b92b76b3b3a4039