input: misc: fpc1145: Add poll for IO multiplex operations.
A poll interface allows applications to wait for a specified timeout
or indefinitely, while at the same time blocking on other file
descriptors such as an eventfd. This is in contrast to the blocking
ioctl which requires a wakeup every so often to check if another event
(such as cancellation) is requested. This wakeup was configured at
approximately .5 seconds, spamming the log and eating useless cpu
cycles.

Change-Id: Ieb34f0d731f257f6789c2d1977435d0d46a1dc4e
Signed-off-by: MarijnS95 <marijns95@gmail.com>